[MSCTFIME] Implement CtfImeSetActiveContextAlways
